feat: return precompile errors as reverts (#473)
ref https://github.com/tempoxyz/tempo/issues/407

This actually implements most of the user-facing changes and succesfully
propagates all precompile errors as reverts to the EVM.

Right now, all errors are mapped to `PrecompileError::Other` variant
which translates to a `PrecompileError` halt. This PR changes this to be
handled as a revert, and correctly propagating the revert reason.

Additionally, invalid calldata is now handled by reverting with empty
calldata which mirrors Solidity behavior.

Open questions:
- Are there cases when we would want the actual halt to happen? halt vs
revert in Solidity terms is `assert` vs `require`
- What should be the gas usage of reverted calls? Halts consume entire
passed gas. Right now I've set gas usage of reverted calls to be the
same as succesful ones but it should probably be lower?