feat(consensus): on chain DKG (#538)
Adds on-chain distributed key generation.

This patch adds two actors to the consensus engine:

1. the epoch manager, that is entering and exiting epochs by
starting/stopping consensus engines for a specific epoch.
2. the dkg manager, that is running DKG ceremonies.

Building on top of https://github.com/tempoxyz/tempo/pull/481, the epoch
manager has been significantly simplified. Because in order to
transition epochs one needs to know the new set of participants
(validator identities) as well as the public polynomial and the private
key share for the new epoch, the epoch manager is now directly slaved to
the dkg manager.

# New config options

Adds a new config option `epoch_length`, the number of heights in an
epoch.

# New metrics and changed metrics

+ `epoch_manager_latest_epoch` that tracks the latest (and thus highest)
epoch for which the orchestrator started a consensus engine.
+ `epoch_manager_active_epochs` that tracks how many consensus engines
are currently active (i.e. how many epochs are backed by an active
engine)
+ `epoch_manager_latest_participants` that tracks how many participants
(i.e. validators) make up the latest epoch
+ `dkg_manager_ceremony_failures` that tracks how many DKG ceremonies
failed
+ `dkg_manager_ceremony_successes` that tracks how many DKG ceremonies
succeeded
+ `dkg_manager_ceremony_dealers` that tracks how many dealers
participate in the active ceremony
+ `dkg_manager_ceremony_players` that tracks how many players
participate in the active ceremony

All metrics relating to the commonware marshal actor have been changed
from `*_syncer_*` to `*_marshal_*`.

# Ceremony details

The DKG manager is running DKG ceremonies in the following way:

1. First, the ceremony is started by creating a new commitment and set
of private key shares.
2. In the first half of the epoch, the DKG manager sends the key shares
to all players for acknowledgment. The DKG manager collects these
acknowledgments.
3. At the epoch midpoint, the DKG manager constructs the outcome of its
dealings (its commitment, the acknowledgments it received, and the key
shares for which it did not receive acks and which it now reveals).
4. During the second half of the epoch, every time the node is asked to
propose, it attempts to fetch the dealing outcomes from its DKG manager
to write it to the `extraData` field of its proposal header.
5. Concurrently, in the second half of the epoch the DKG manager reads
these dealing outcomes from the `extraData` field of each block header.
6. If the DKG manager finds its own dealing outcome in a block, it
removes it so that the node does not write it to a block again (note
that there is a chance that it finds the same dealing outcomes twice
because of finalization delays).
7. On the pre-to-last block of an epoch, the DKG manager generates the
overall outcome of the ceremony so it can be written to the last block
of the epoch. It also already starts the ceremony for the next epoch
here and requests the epoch manager to start a new epoch.
8. If a node is the proposer on the last block of the epoch, it writes
this ceremony outcome to the `extraData` header of its proposal. If the
ceremony outcome is not yet ready, it refuses to propose.
9. If a node is the validator on the last block of the epoch, it
verifies that the last block indeed contains the same ceremony outcome.