fix(consensus): relax starting conditions on height mismatch between execution, consensus layers (#768)
The starting conditions were too strict. It's probably fine if our
execution layer is ahead of the consensus layer. We just need to set the
forkchoice state to whatever is higher. Consensus will then forward one
block or more blocks, and execution will process or drop it.

This change is not breaking.

Followup: https://github.com/tempoxyz/tempo/issues/769